First revision of the model viewer prototype. cmake not working, feel free to change it. Working using qmake.
The loader is able to load the obj file from model.cpp. For various
reasons, the sourcecode is disgusting --> the program was randomly
crashing but that's just a prototype.  Use the zqsd and ijklpm to move
the model. The mouse part is not working. Weird render with bigger
models, most likely due to the projection parameters.